Entrega3/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.13
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
scipy>=1.11.0
